Remove dependency on inventory
Wrap `C_GetFunctionList` with Backend registration logic instead.
