+++ native-pkcs11-traits/Cargo.toml
@@ -11,7 +11,6 @@ repository = "https://github.com/google/native-pkcs11"
 license = "Apache-2.0"
 
 [dependencies]
-inventory = "0.3.3"
 once_cell = "1.17.0"
 rand = "0.8.5"
 x509-cert = { version = "0.1.1", default-features = false }
