+++ native-pkcs11/Cargo.toml
@@ -10,8 +10,10 @@ rust-version = "1.63"
 repository = "https://github.com/google/native-pkcs11"
 license = "Apache-2.0"
 
+[features]
+custom-function-list = []
+
 [dependencies]
-inventory = "0.3.3"
 native-pkcs11-core = { path = "../native-pkcs11-core" }
 native-pkcs11-traits = { version = "0.1.0", path = "../native-pkcs11-traits" }
 once_cell = "1.17.0"
