+++ pkcs11-keychain/src/macos/mod.rs
@@ -16,17 +16,13 @@ use std::fmt::Debug;
 
 pub use backend::KeychainBackend;
 use core_foundation::error::CFError;
-use native_pkcs11_traits::{once_cell::sync::Lazy, Backend, RegisteredBackend};
+use native_pkcs11_traits::{once_cell::sync::Lazy, Backend};
 use thiserror::Error;
 use tracing_error::SpanTrace;
 
 mod backend;
 pub mod certificate;
 pub mod key;
-
-static BACKEND: Lazy<Box<dyn Backend>> = Lazy::new(|| Box::new(KeychainBackend::new()));
-
-native_pkcs11_traits::inventory::submit! {RegisteredBackend(&BACKEND)}
 
 //  NOTE(kcking): I think this just works because any non-System path defaults
 //  to the Login keychain
